% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_table.R
\name{kable_with_scroll_box}
\alias{kable_with_scroll_box}
\title{Print table with scroll box}
\usage{
kable_with_scroll_box(
  table,
  scroll.height = "300px",
  scroll.width = "100\%",
  scroll.fixed_thead = TRUE,
  ...
)
}
\arguments{
\item{table}{A input table}

\item{scroll.height}{The height param in scroll_box()}

\item{scroll.width}{The width param in scroll_box()}

\item{scroll.fixed_thead}{The fixed_thead param in scroll_box()}

\item{...}{Arguments from kableExtra::kable}
}
\value{
A styling table with scroll box
}
\description{
Print table with scroll box
}
\examples{
kable_with_scroll_box(iris)

}
