% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_table.R
\name{kable_with_style}
\alias{kable_with_style}
\title{Print table with the style}
\usage{
kable_with_style(table, ...)
}
\arguments{
\item{table}{A input table}

\item{...}{Arguments from kableExtra::kable}
}
\value{
A styling table
}
\description{
Print the input table with following styles, bootstrap_options = c("striped", "hover", "condensed", "responsive")
}
\examples{
kable_with_style(iris)
}
