% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wesbox-package.R
\docType{package}
\name{wesbox-package}
\alias{wesbox}
\alias{wesbox-package}
\title{wesbox: The toolBOX for the Whole-Exome Sequencing data analysis}
\description{
A collection of functions used in my daily data analysis.
}
\author{
\strong{Maintainer}: Jinjie Duan \email{jjduan@biomed.au.dk} (\href{https://orcid.org/0000-0002-3402-302X}{ORCID})

}
\keyword{internal}
